♻️ refactor(dashboard): move dashboard into feature module
Move dashboard components, styles, and data hooks under src/features/dashboard while keeping the route page as a thin export.

This establishes the first vertical feature slice without changing dashboard routing or UI behavior.
