+++ src/features/dashboard/DashboardPage.module.scss
@@ -1,6 +1,6 @@
 @use 'sass:color';
-@use '../styles/variables.scss' as *;
-@use '../styles/mixins.scss' as *;
+@use '../../styles/variables.scss' as *;
+@use '../../styles/mixins.scss' as *;
 
 // ─── Container ──────────────────────────────────────────
