+++ src/features/dashboard/DashboardPage.tsx
@@ -1,0 +1,436 @@
+import { useCallback, useEffect, useRef, useState } from 'react';
+import { Link } from 'react-router-dom';
+import { useTranslation } from 'react-i18next';
+import { IconKey, IconBot, IconFileText, IconSatellite } from '@/components/ui/icons';
+import { useAuthStore, useConfigStore, useModelsStore } from '@/stores';
+import { apiKeysApi, providersApi, authFilesApi } from '@/services/api';
+import { useHeaderRefresh } from '@/hooks/useHeaderRefresh';
+import { VersionCard } from './components/VersionCard';
+import { UsageMetricsCard } from './components/UsageMetricsCard';
+import { CollectorStatusCard } from './components/CollectorStatusCard';
+import { HealthAlertsCard } from './components/HealthAlertsCard';
+import { TrafficOverviewCard } from './components/TrafficOverviewCard';
+import { useDashboardUsageSummary } from './hooks/useDashboardUsageSummary';
+import styles from './DashboardPage.module.scss';
+
+interface QuickStat {
+  label: string;
+  value: number | string;
+  icon: React.ReactNode;
+  path: string;
+  loading?: boolean;
+  sublabel?: string;
+}
+
+interface ProviderStats {
+  gemini: number | null;
+  codex: number | null;
+  claude: number | null;
+  openai: number | null;
+}
+
+export function DashboardPage() {
+  const { t, i18n } = useTranslation();
+  const connectionStatus = useAuthStore((state) => state.connectionStatus);
+  const serverVersion = useAuthStore((state) => state.serverVersion);
+  const serverBuildDate = useAuthStore((state) => state.serverBuildDate);
+  const apiBase = useAuthStore((state) => state.apiBase);
+  const managementKey = useAuthStore((state) => state.managementKey);
+  const config = useConfigStore((state) => state.config);
+  const usageSummary = useDashboardUsageSummary();
+  const refreshUsageSummary = usageSummary.refresh;
+
+  const models = useModelsStore((state) => state.models);
+  const modelsLoading = useModelsStore((state) => state.loading);
+  const fetchModelsFromStore = useModelsStore((state) => state.fetchModels);
+
+  const [stats, setStats] = useState<{
+    apiKeys: number | null;
+    authFiles: number | null;
+  }>({
+    apiKeys: null,
+    authFiles: null,
+  });
+
+  const [providerStats, setProviderStats] = useState<ProviderStats>({
+    gemini: null,
+    codex: null,
+    claude: null,
+    openai: null,
+  });
+
+  const [loading, setLoading] = useState(true);
+
+  const [currentTime, setCurrentTime] = useState(() => new Date());
+  const [cardRefreshSignal, setCardRefreshSignal] = useState(0);
+
+  const apiKeysCache = useRef<string[]>([]);
+
+  useEffect(() => {
+    apiKeysCache.current = [];
+  }, [apiBase, config?.apiKeys]);
+
+  // Update time every 60 seconds
+  useEffect(() => {
+    const id = setInterval(() => {
+      setCurrentTime(new Date());
+    }, 60_000);
+    return () => clearInterval(id);
+  }, []);
+
+  const normalizeApiKeyList = (input: unknown): string[] => {
+    if (!Array.isArray(input)) return [];
+    const seen = new Set<string>();
+    const keys: string[] = [];
+
+    input.forEach((item) => {
+      const record =
+        item !== null && typeof item === 'object' && !Array.isArray(item)
+          ? (item as Record<string, unknown>)
+          : null;
+      const value =
+        typeof item === 'string'
+          ? item
+          : record
+            ? (record['api-key'] ?? record['apiKey'] ?? record.key ?? record.Key)
+            : '';
+      const trimmed = String(value ?? '').trim();
+      if (!trimmed || seen.has(trimmed)) return;
+      seen.add(trimmed);
+      keys.push(trimmed);
+    });
+
+    return keys;
+  };
+
+  const resolveApiKeysForModels = useCallback(async () => {
+    if (apiKeysCache.current.length) {
+      return apiKeysCache.current;
+    }
+
+    const configKeys = normalizeApiKeyList(config?.apiKeys);
+    if (configKeys.length) {
+      apiKeysCache.current = configKeys;
+      return configKeys;
+    }
+
+    try {
+      const list = await apiKeysApi.list();
+      const normalized = normalizeApiKeyList(list);
+      if (normalized.length) {
+        apiKeysCache.current = normalized;
+      }
+      return normalized;
+    } catch {
+      return [];
+    }
+  }, [config?.apiKeys]);
+
+  const fetchModels = useCallback(async () => {
+    if (connectionStatus !== 'connected' || !apiBase) {
+      return;
+    }
+
+    try {
+      const apiKeys = await resolveApiKeysForModels();
+      const primaryKey = apiKeys[0];
+      await fetchModelsFromStore(apiBase, primaryKey);
+    } catch {
+      // Ignore model fetch errors on dashboard
+    }
+  }, [connectionStatus, apiBase, resolveApiKeysForModels, fetchModelsFromStore]);
+
+  const refreshStats = useCallback(async () => {
+    if (connectionStatus === 'connected') {
+      setLoading(true);
+      try {
+        const [keysRes, filesRes, geminiRes, codexRes, claudeRes, openaiRes] =
+          await Promise.allSettled([
+            apiKeysApi.list(),
+            authFilesApi.list(),
+            providersApi.getGeminiKeys(),
+            providersApi.getCodexConfigs(),
+            providersApi.getClaudeConfigs(),
+            providersApi.getOpenAIProviders(),
+          ]);
+
+        setStats({
+          apiKeys: keysRes.status === 'fulfilled' ? keysRes.value.length : null,
+          authFiles: filesRes.status === 'fulfilled' ? filesRes.value.files.length : null,
+        });
+
+        setProviderStats({
+          gemini: geminiRes.status === 'fulfilled' ? geminiRes.value.length : null,
+          codex: codexRes.status === 'fulfilled' ? codexRes.value.length : null,
+          claude: claudeRes.status === 'fulfilled' ? claudeRes.value.length : null,
+          openai: openaiRes.status === 'fulfilled' ? openaiRes.value.length : null,
+        });
+      } finally {
+        setLoading(false);
+      }
+    } else {
+      setLoading(false);
+    }
+  }, [connectionStatus]);
+
+  const refreshDashboard = useCallback(async () => {
+    setCurrentTime(new Date());
+    setCardRefreshSignal((value) => value + 1);
+    await Promise.all([refreshStats(), fetchModels(), refreshUsageSummary()]);
+  }, [fetchModels, refreshStats, refreshUsageSummary]);
+
+  useEffect(() => {
+    void Promise.all([refreshStats(), fetchModels()]);
+  }, [fetchModels, refreshStats]);
+
+  useHeaderRefresh(refreshDashboard);
+
+  // Calculate total provider keys only when all provider stats are available.
+  const providerStatsReady =
+    providerStats.gemini !== null &&
+    providerStats.codex !== null &&
+    providerStats.claude !== null &&
+    providerStats.openai !== null;
+  const hasProviderStats =
+    providerStats.gemini !== null ||
+    providerStats.codex !== null ||
+    providerStats.claude !== null ||
+    providerStats.openai !== null;
+  const totalProviderKeys = providerStatsReady
+    ? (providerStats.gemini ?? 0) +
+      (providerStats.codex ?? 0) +
+      (providerStats.claude ?? 0) +
+      (providerStats.openai ?? 0)
+    : 0;
+
+  const quickStats: QuickStat[] = [
+    {
+      label: t('dashboard.management_keys'),
+      value: stats.apiKeys ?? '-',
+      icon: <IconKey size={24} />,
+      path: '/config',
+      loading: loading && stats.apiKeys === null,
+      sublabel: t('nav.config_management'),
+    },
+    {
+      label: t('nav.ai_providers'),
+      value: loading ? '-' : providerStatsReady ? totalProviderKeys : '-',
+      icon: <IconBot size={24} />,
+      path: '/ai-providers',
+      loading: loading,
+      sublabel: hasProviderStats
+        ? t('dashboard.provider_keys_detail', {
+            gemini: providerStats.gemini ?? '-',
+            codex: providerStats.codex ?? '-',
+            claude: providerStats.claude ?? '-',
+            openai: providerStats.openai ?? '-',
+          })
+        : undefined,
+    },
+    {
+      label: t('nav.auth_files'),
+      value: stats.authFiles ?? '-',
+      icon: <IconFileText size={24} />,
+      path: '/auth-files',
+      loading: loading && stats.authFiles === null,
+      sublabel: t('dashboard.oauth_credentials'),
+    },
+    {
+      label: t('dashboard.available_models'),
+      value: modelsLoading ? '-' : models.length,
+      icon: <IconSatellite size={24} />,
+      path: '/system',
+      loading: modelsLoading,
+      sublabel: t('dashboard.available_models_desc'),
+    },
+  ];
+
+  const routingStrategyRaw = config?.routingStrategy?.trim() || '';
+  const routingStrategyDisplay = !routingStrategyRaw
+    ? '-'
+    : routingStrategyRaw === 'round-robin'
+      ? t('basic_settings.routing_strategy_round_robin')
+      : routingStrategyRaw === 'fill-first'
+        ? t('basic_settings.routing_strategy_fill_first')
+        : routingStrategyRaw;
+  const routingStrategyBadgeClass = !routingStrategyRaw
+    ? styles.configBadgeUnknown
+    : routingStrategyRaw === 'round-robin'
+      ? styles.configBadgeRoundRobin
+      : routingStrategyRaw === 'fill-first'
+        ? styles.configBadgeFillFirst
+        : styles.configBadgeUnknown;
+
+  const formattedDate = currentTime.toLocaleDateString(i18n.language, {
+    weekday: 'long',
+    year: 'numeric',
+    month: 'long',
+    day: 'numeric',
+  });
+
+  const formattedTime = currentTime.toLocaleTimeString(i18n.language, {
+    hour: '2-digit',
+    minute: '2-digit',
+  });
+
+  return (
+    <div className={styles.dashboard}>
+      <section className={styles.hero}>
+        <div className={styles.heroMeta}>
+          <div className={styles.dateTimeBlock}>
+            <span className={styles.time}>{formattedTime}</span>
+            <span className={styles.date}>{formattedDate}</span>
+          </div>
+          <div className={styles.connectionPill}>
+            <span
+              className={`${styles.statusDot} ${
+                connectionStatus === 'connected'
+                  ? styles.connected
+                  : connectionStatus === 'connecting'
+                    ? styles.connecting
+                    : styles.disconnected
+              }`}
+            />
+            <span className={styles.pillText}>
+              {t(
+                connectionStatus === 'connected'
+                  ? 'common.connected'
+                  : connectionStatus === 'connecting'
+                    ? 'common.connecting'
+                    : 'common.disconnected'
+              )}
+            </span>
+          </div>
+        </div>
+      </section>
+
+      <VersionCard
+        appVersion={__APP_VERSION__ || t('dashboard.version_unknown')}
+        apiVersion={serverVersion || t('dashboard.version_unknown')}
+        apiBase={apiBase || ''}
+        serverBuildDate={serverBuildDate || undefined}
+        connectionStatus={connectionStatus}
+        refreshSignal={cardRefreshSignal}
+      />
+
+      {usageSummary.enabled && (
+        <section className={styles.usageSection}>
+          <div className={styles.usageCardGrid}>
+            <div className={styles.trafficOverview}>
+              <TrafficOverviewCard
+                timeline={usageSummary.trafficTimeline}
+                hourlyActivity={usageSummary.hourlyActivity}
+                tokenMix={usageSummary.tokenMix}
+                loading={usageSummary.loading}
+              />
+            </div>
+            <UsageMetricsCard
+              summary={usageSummary.summary}
+              topModels={usageSummary.topModels}
+              modelCostRank={usageSummary.modelCostRank}
+              loading={usageSummary.loading}
+              error={usageSummary.error}
+              lastRefreshedAt={usageSummary.lastRefreshedAt}
+            />
+            <CollectorStatusCard
+              enabled={usageSummary.enabled}
+              serviceBase={usageSummary.serviceBase}
+              managementKey={managementKey}
+              refreshSignal={cardRefreshSignal}
+            />
+            <HealthAlertsCard
+              enabled={usageSummary.enabled}
+              loading={usageSummary.loading}
+              recentFailures={usageSummary.recentFailures}
+              channelHealth={usageSummary.channelHealth}
+              failureSources={usageSummary.failureSources}
+              refreshSignal={cardRefreshSignal}
+            />
+          </div>
+        </section>
+      )}
+
+      {/* Bento stats grid */}
+      <section className={styles.statsSection}>
+        <h2 className={styles.sectionHeading}>{t('dashboard.system_overview')}</h2>
+        <div className={styles.bentoGrid}>
+          {quickStats.map((stat, index) => (
+            <Link
+              key={stat.path}
+              to={stat.path}
+              className={`${styles.bentoCard} ${index === 0 ? styles.bentoLarge : ''}`}
+              style={{ animationDelay: `${index * 80}ms` }}
+            >
+              <div className={styles.bentoIcon}>{stat.icon}</div>
+              <div className={styles.bentoContent}>
+                <span className={styles.bentoValue}>{stat.loading ? '...' : stat.value}</span>
+                <span className={styles.bentoLabel}>{stat.label}</span>
+                {stat.sublabel && !stat.loading && (
+                  <span className={styles.bentoSublabel}>{stat.sublabel}</span>
+                )}
+              </div>
+            </Link>
+          ))}
+        </div>
+      </section>
+
+      {/* Config pills section */}
+      {config && (
+        <section className={styles.configSection}>
+          <h2 className={styles.sectionHeading}>{t('dashboard.current_config')}</h2>
+          <div className={styles.configPillGrid}>
+            <div className={styles.configPill}>
+              <span className={styles.configPillLabel}>{t('basic_settings.debug_enable')}</span>
+              <span
+                className={`${styles.configPillValue} ${config.debug ? styles.on : styles.off}`}
+              >
+                {config.debug ? t('common.yes') : t('common.no')}
+              </span>
+            </div>
+            <div className={styles.configPill}>
+              <span className={styles.configPillLabel}>
+                {t('basic_settings.logging_to_file_enable')}
+              </span>
+              <span
+                className={`${styles.configPillValue} ${config.loggingToFile ? styles.on : styles.off}`}
+              >
+                {config.loggingToFile ? t('common.yes') : t('common.no')}
+              </span>
+            </div>
+            <div className={styles.configPill}>
+              <span className={styles.configPillLabel}>
+                {t('basic_settings.retry_count_label')}
+              </span>
+              <span className={styles.configPillValue}>{config.requestRetry ?? 0}</span>
+            </div>
+            <div className={styles.configPill}>
+              <span className={styles.configPillLabel}>{t('basic_settings.ws_auth_enable')}</span>
+              <span
+                className={`${styles.configPillValue} ${config.wsAuth ? styles.on : styles.off}`}
+              >
+                {config.wsAuth ? t('common.yes') : t('common.no')}
+              </span>
+            </div>
+            <div className={styles.configPill}>
+              <span className={styles.configPillLabel}>{t('dashboard.routing_strategy')}</span>
+              <span className={`${styles.configBadge} ${routingStrategyBadgeClass}`}>
+                {routingStrategyDisplay}
+              </span>
+            </div>
+            {config.proxyUrl && (
+              <div className={`${styles.configPill} ${styles.configPillWide}`}>
+                <span className={styles.configPillLabel}>
+                  {t('basic_settings.proxy_url_label')}
+                </span>
+                <span className={styles.configPillMono}>{config.proxyUrl}</span>
+              </div>
+            )}
+          </div>
+          <Link to="/config" className={styles.viewMoreLink}>
+            {t('dashboard.edit_settings')} →
+          </Link>
+        </section>
+      )}
+    </div>
+  );
+}
