+++ src/features/dashboard/components/CollectorStatusCard.module.scss
