+++ src/features/dashboard/components/CollectorStatusCard.tsx
