+++ src/features/dashboard/components/HealthAlertsCard.module.scss
